+++ packages/server/src/modules/posts/posts.controller.ts
@@ -12,7 +12,7 @@ import {
 import { PostsService } from './posts.service';
 import { CreatePostDto } from './dto/create-post.dto';
 import { GetCurrentUser, Public } from '~/common/decorators';
-import { ApiBody, ApiConsumes, ApiTags } from '@nestjs/swagger';
+import { ApiTags } from '@nestjs/swagger';
 import { FileInterceptor } from '@nestjs/platform-express';
 import multerOptions from '~/utils/multerOptions';
 import { GetPostsQueryDto } from './dto/get-post-query.dto';
@@ -50,28 +50,12 @@ export class PostsController {
   }
 
   @Post()
-  @ApiConsumes('multipart/form-data')
-  @ApiBody({
-    schema: {
-      type: 'object',
-      properties: {
-        title: { type: 'string' },
-        body: { type: 'string' },
-        tags: { type: 'array', items: { type: 'string' } },
-        file: {
-          type: 'string',
-          format: 'binary',
-        },
-      },
-    },
-  })
   @UseInterceptors(FileInterceptor('file', multerOptions))
   async createPost(
     @Body() dto: CreatePostDto,
     @GetCurrentUser('userId') userId: string,
-    @UploadedFile() file?: Express.Multer.File,
   ) {
-    return await this.postsService.createPost(dto, userId, file);
+    return await this.postsService.createPost(dto, userId);
   }
 
   @Post(':id/likes')
@@ -96,5 +80,11 @@ export class PostsController {
     @GetCurrentUser('userId') userId: string,
   ): Promise<void> {
     return await this.postsService.deletePost(id, userId);
+  }
+
+  @Post('upload')
+  @UseInterceptors(FileInterceptor('file', multerOptions))
+  async uploadImage(@UploadedFile() file: Express.Multer.File) {
+    return await this.postsService.uploadImage(file);
   }
 }
