+++ packages/server/src/modules/posts/posts.service.ts
@@ -1,5 +1,7 @@
 import { HttpException, Injectable } from '@nestjs/common';
+import { ConfigService } from '@nestjs/config';
 import { Post, PostLike } from '@prisma/client';
+import { S3Service } from '~/providers/aws/s3/s3.service';
 import { CommentsService } from '../comments/comments.service';
 import { CreatePostDto } from './dto/create-post.dto';
 import { PostsRepository } from './posts.repository';
@@ -9,6 +11,8 @@ export class PostsService {
   constructor(
     private readonly postRepository: PostsRepository,
     private readonly commentsService: CommentsService,
+    private readonly configService: ConfigService,
+    private readonly s3Service: S3Service,
   ) {}
   async getPosts(cursor: string | null, userId: string | null) {
     const { totalCount, endCursor, hasNextPage, list } =
@@ -32,12 +36,8 @@ export class PostsService {
     return this.serializePost(post);
   }
 
-  async createPost(
-    dto: CreatePostDto,
-    userId: string,
-    file?: Express.Multer.File,
-  ) {
-    const post = await this.postRepository.createPost(dto, userId, file);
+  async createPost(dto: CreatePostDto, userId: string) {
+    const post = await this.postRepository.createPost(dto, userId);
     const postStats = await this.postRepository.createPostStats(post.id);
 
     const postWithStats = { ...post, postStats };
@@ -97,11 +97,27 @@ export class PostsService {
     if (post.userId !== userId)
       throw new HttpException('You are not the author of this post', 403);
 
-    await this.postRepository.deleteImage(post.thumbnail);
+    await this.deleteImage(post.thumbnail);
     await this.postRepository.deletePost(id);
   }
 
   async getPostComments(id: string, userId: string | null) {
     return await this.commentsService.getComments(id, userId);
+  }
+
+  async uploadImage(file: Express.Multer.File) {
+    const filename = `${Date.now()}-${file.originalname}`;
+    await this.s3Service.pubObject(filename, file);
+    return filename;
+  }
+
+  async deleteImage(thumbnail: string) {
+    const bucket = this.configService.get<string>('AWS_S3_BUCKET');
+    const filename = thumbnail.replace(
+      `https://${bucket}.s3.amazonaws.com/`,
+      '',
+    );
+    await this.s3Service.deleteObject(filename);
+    return filename;
   }
 }
